+++ NEWS.md
@@ -4,7 +4,8 @@
 * Added ActArray#power? method.
 * Added #beams and #stored attributes for Gripper class.
 * Added #position, #speed, #acceleration and #current attributes for Actor class.
-* Added #px, #py, #pa, #vx, #vy, #va and #power? attributes for Actor class. Method #power is deprecated.
+* Added #px, #py, #pa, #vx, #vy, #va and #power? attributes for Position2d class. Method #power is deprecated.
+* Added #volts, #percent, #joules and #watts attributes for Power class.
 
 ## 2012-03-06 0.4.1